% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manip.R
\name{drop_vars}
\alias{drop_vars}
\alias{mapped_vars}
\alias{data_vars}
\title{Drop unused variables from data}
\usage{
drop_vars(p, keep.vars = character(), guess.vars = TRUE)

mapped_vars(p, invert = FALSE)

data_vars(p)
}
\arguments{
\item{p}{ggplot Plot object with embedded data.}

\item{keep.vars}{character Names of unused variables to be kept.}

\item{guess.vars}{logical Flag indicating whether to find used variables
automatically.}

\item{invert}{logical If TRUE return indices for elements of \code{data} that
are not mapped to any aesthetic or facet.}
}
\value{
character vector with names of mapped variables in the default
   data object.

character vector with names of mapped variables in the default
   data object.
}
\description{
Automatically remove unused variables from the data object embedded in a
\code{gg} or \code{ggplot} object.
}
\note{
This function is under development and not yet thoroughly tested! It is
  a demonstration of how one can manipulate the internals of \code{ggplot}
  objects. This function may stop working after some future update to the
  'ggplot2' package.

  Rather than using this function after creating the \code{ggplot} object it
  may be more efficient to extract the variables of interest and pass a data
  frame containing only these to the \code{ggplot()} constructor.
}
\section{Warning!}{
 The current implementation drops variables only from the
  default data object. Data objects within layers are not modified.
}

\examples{
library(ggplot2)

p <- ggplot(mpg, aes(factor(year), (cty + hwy) / 2)) +
  geom_boxplot() +
  facet_grid(. ~ class)

mapped_vars(p) # those in use
mapped_vars(p, invert = TRUE) # those not used

p.dp <- drop_vars(p) # we drop unused vars

# number of columns in the data member
ncol(p$data)
ncol(p.dp$data)

# which vars are in the data member
data_vars(p)
data_vars(p.dp)

# which variables in data are used in the plot
mapped_vars(p)
mapped_vars(p.dp)

# the plots identical
p
p.dp

# structure and size of p
str(p, max.level = 0)
str(p.dp, max.level = 0) # smaller in size

# structure and size of p["data"]
str(p, components = "data")
str(p.dp, components = "data") # smaller in size

}
