% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-debug.R, R/stat-debug-group.R,
%   R/stat-debug-panel.R
\docType{data}
\name{GeomNull}
\alias{GeomNull}
\alias{GeomDebug}
\alias{StatDebugGroup}
\alias{StatDebugPanel}
\title{'ggproto' objects}
\description{
Internal use only definitions of 'ggproto' objects needed for
  geoms and stats.
}
\keyword{datasets}
\keyword{internal}
