% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gginnards.R
\docType{package}
\name{gginnards-package}
\alias{gginnards}
\alias{gginnards-package}
\title{gginnards: Explore the Innards of 'ggplot'2' Objects}
\description{
Extensions to 'ggplot2' providing low-level debug tools: statistics
and geometries reporting data passed to compute_group() and compute_panel()
functions and to geometries. Layer manipulation: funtions for deletion,
insertion, extraction and reordering of layers of ggplot objects.
}
\details{
The new facilities for cleanly defining new stats and geoms added to
'ggplot2' in version 2.0.0 have made this package easy to code. However, this
means that this package requires version 2.0.0 or later of ggplot2.

Extensions provided:
\itemize{
\item "Debug" stats and a "debug" geom that print to the console a summary
of their \code{data} input.
\item Functions for inspecting and manipulating the layers of a ggplot
object.
}
}
\references{
Package 'ggplot2' web site at \url{http://ggplot2.org/}\cr
Package 'ggplot2' documentation at \url{http://docs.ggplot2.org/}\cr
Package 'ggplot2' source code at \url{https://github.com/hadley/ggplot2}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.r4photobiology.info}
  \item \url{https://bitbucket.org/aphalo/gginnards}
  \item Report bugs at \url{https://bitbucket.org/aphalo/gginnards/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (0000-0003-3385-972X)

}
