% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-inspection.R
\name{theme_diff}
\alias{theme_diff}
\alias{theme_classes}
\alias{theme_extract}
\title{Theme difference}
\usage{
theme_diff(
  theme1,
  theme2 = ggplot2::theme_grey(),
  pattern = NULL,
  classes = NULL,
  value = FALSE,
  invert = FALSE
)

theme_classes(theme1, pattern = NULL)

theme_extract(theme1, pattern = NULL, classes = NULL, value = TRUE)
}
\arguments{
\item{theme1, theme2}{'ggplot2' theme objects.}

\item{pattern}{character A character string containing a regular expression
to be matched by \code{\link{grep}()} against the union of the field
names in to two themes being compared or in the theme being operated
upon. If \code{NULL} all fields are used.}

\item{classes}{character A character vector containg the names of classes
of the fields to operate upon. For ggplot2 elements, without "ggplot2::".}

\item{value}{logical If \code{TRUE} return a list with the fields with
differences extracted from the themes, and if \code{FALSE}, return the
names of the fields with differences.}

\item{invert}{logical If \code{TRUE} return the fields that are equal instead
of those that are different.}
}
\value{
The value returned by \code{theme_diff()} is a \code{character}
  vector of field names if \code{value = FALSE} or otherwise a named list
  with two members, each a subset of one of the themes, named
  \code{"theme1"} and  \code{"theme2"}. The value returned by
  \code{theme_classes()} is a named \code{character}
  vector of most derived class names, with field names as names.
  The value returned by \code{theme_extract()} is a \code{character}
  vector of field names if \code{value = FALSE} or otherwise a subset of
  the theme fields.
}
\description{
Compare two 'ggplot2' theme definitions, field by field.

Extract the class of the members of a 'ggplot2' theme.
}
\details{
The names of the shared fields obtained with \code{\link{setdiff}()}
  are used to walk through the list-like theme definitions, comparing each of
  these fields with \code{\link{identical}()}. Fields missing or \code{NULL}
  in only one of the two themes, are also considered different. If
  \code{pattern} is not \code{NULL}, the comparison is retricted to the
  field names matching the pattern. If \code{classes} is not \code{NULL},
  the comparison is restricted to the fields of a class listed in the
  vector passed as argument.
}
\examples{

theme_diff(theme_bw())
theme_diff(theme_bw(), theme_grey())

theme_diff(theme_bw(), theme_bw(base_size = 12))
theme_diff(theme_bw(), theme_bw(base_size = 12), pattern = "^strip")
theme_diff(theme_bw(), theme_bw(base_size = 12), classes = "element_text")
theme_diff(theme_bw(), theme_bw(base_size = 12), classes = "simpleUnit")

theme_diff(theme_classic(), theme_bw())
theme_diff(theme_classic(), theme_bw(), pattern = "^axis\\\\.title", invert = TRUE)
theme_diff(theme_bw(), theme_bw(base_size = 12), pattern = "title$", value = TRUE)
theme_diff(theme_bw(), theme_bw(base_size = 12), classes = "element_rect", value = TRUE)
theme_diff(theme_bw(), theme_bw(base_size = 12), classes = "simpleUnit", value = TRUE)

theme_classes(theme_bw(), pattern = "^text$")

theme_diff(theme_bw(), theme_bw(base_size = 12), pattern = "^text$", value = TRUE)

theme_extract(theme_bw(), pattern = "^text$")
theme_extract(theme_bw(12), pattern = "^text$")

# fields that are not set explicitly (inherited)
theme_extract(theme_bw(), classes = "NULL", value = FALSE)


theme_classes(theme_bw(), pattern = "border")


theme_extract(theme_bw(), pattern = "border")

}
