% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-debug.R, R/stat-debug.R
\docType{data}
\name{GeomNull}
\alias{GeomNull}
\alias{debug_draw_function}
\alias{GeomDebugPanel}
\alias{GeomDebugGroup}
\alias{debug_compute_fun}
\alias{StatDebugGroup}
\alias{StatDebugPanel}
\title{'ggproto' objects}
\description{
Internal use only definitions of 'ggproto' objects needed for
  geoms and stats.
}
\keyword{datasets}
\keyword{internal}
