% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-diff.R
\name{theme_diff}
\alias{theme_diff}
\alias{theme_classes}
\title{Theme difference}
\usage{
theme_diff(
  theme1,
  theme2 = ggplot2::theme_gray(),
  pattern = NULL,
  value = FALSE,
  invert = FALSE
)

theme_classes(theme1, pattern = NULL)
}
\arguments{
\item{theme1, theme2}{'ggplot2' theme objects.}

\item{pattern}{character A character string containing a regular expression
to be matched by \code{\link{grep}()} against the union of the field
names in to two themes being compared or in the theme being operated
upon. If \code{NULL} all fields are used.}

\item{value}{logical If \code{TRUE} return a list with the fields with
differences extracted from the themes, and if \code{FALSE}, return the
names of the fields with differences.}

\item{invert}{logical If \code{TRUE} return the fields that are equal instead
of those that are different.}
}
\value{
A \code{character} vector if \code{value = FALSE} or otherwise a
  list with two members, each a subset of one of the themes, named
  \code{"theme1"} and  \code{"theme2"}.
}
\description{
Compare two 'ggplot2' theme definitions, field by field.

Extract the class of the members of a 'ggplot2'
}
\details{
The names of the shared fields obtained with \code{\link{setdiff}()}
  are used to walk through the list-like theme definitions, comparing each of
  these fields with \code{\link{all.equal}()}. Fields missing from either of
  the two themes, are also considered different.
}
\examples{

theme_diff(theme_bw())
theme_diff(theme_bw(), theme_gray())

theme_diff(theme_bw(), theme_bw(base_size = 12))
theme_diff(theme_bw(), theme_bw(base_size = 12), pattern = "^strip")

theme_diff(theme_classic(), theme_bw())
theme_diff(theme_classic(), theme_bw(), pattern = "^axis\\\\.title", invert = TRUE)
theme_diff(theme_bw(), theme_bw(base_size = 12), value = TRUE)


theme_classes(theme_bw(), pattern = "border")

}
